% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotBCR.R
\name{all_communtiy_algorithms}
\alias{all_communtiy_algorithms}
\title{Provides all community algorithms available}
\usage{
all_communtiy_algorithms()
}
\description{
\code{all_communtiy_algorithms} provides an easy dictionary of all algorithms to determine the communities. This enables access to a human readable representation of the internal functions. "Foo Bar" could be called foo_bar internally, which is not something supposed to being displayed. It will returns a dictionary mapping from \code{string} to \code{string}. The \code{plot_graph} will handle the decryption of the functions internally.
}
\examples{
all_communtiy_algorithms()

}
\seealso{
\code{\link[igraph]{igraph}}

\code{\link[igraph]{communities}}
}
\keyword{algorithm}
\keyword{community}
\keyword{igraph}
\keyword{provider}

