% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceMetric.R
\name{all_distance_metrics}
\alias{all_distance_metrics}
\title{Distance metric shortform mapping}
\usage{
all_distance_metrics()
}
\description{
\code{all_distance_metrics} provides easy access to all distance metric algorithms.
}
\keyword{all_distance_metrics}

