% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotBCR.R
\name{all_layout_algorithms}
\alias{all_layout_algorithms}
\title{Provides all layout algorithms}
\usage{
all_layout_algorithms()
}
\description{
\code{all_layout_algorithms} provides an easy dictionary of all algorithms to determine the layouts. This enables access to a human readable representation of the internal functions. "Foo Bar" could be called foo_bar internally, which is not something supposed to being displayed. It will returns a dictionary mapping from \code{string} to \code{string}. The \code{plot_graph} will handle the decryption of the functions internally.
}
\seealso{
\code{\link[igraph]{igraph}}
}
\keyword{all_layout_algorithms}

