% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceMetric.R
\name{distanceb2b}
\alias{distanceb2b}
\alias{DM}
\title{calculate distance between two BCRs.}
\usage{
distanceb2b(bcr1, bcr2)
}
\arguments{
\item{bcr1}{First BCR}

\item{bcr2}{Second BCR}
}
\value{
Distance between first and second BCR determined by Damerau-Levenshtein Distance
}
\description{
\code{distanceb2b} returns the Damerau-Levenshtein Distance of its arguments.
}
\details{
More Information here.
}
\examples{
distanceb2b("", "")
distanceb2b("ATG", "ATC")

}
\seealso{
\code{\link[stringdist]{stringdist}}
}
\keyword{bcr}
\keyword{damerau}
\keyword{distance}
\keyword{levenshtein}
\keyword{metric}
