% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildIGraph.R
\name{getMapOfBcrs}
\alias{getMapOfBcrs}
\title{returns an environments contain the bcrs and its number of occurrence}
\usage{
getMapOfBcrs(arrayBcr)
}
\arguments{
\item{arrayBcr}{Array of bcrs, should contain only unique values}
}
\description{
returns an environments contain the bcrs and its number of occurrence
}

