% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildIGraph.R
\name{calculateDistances}
\alias{BIG}
\alias{calculateDistances}
\title{calculate distances between given bcr arrays}
\usage{
calculateDistances(arrayBcr, distanceMetric = "LD", parameter = -1,
  nthread = -1)
}
\arguments{
\item{arrayBcr}{Array of bcrs, should contain only unique values}

\item{distanceMetric}{Possilble distance metrics ("LD" - Levenshtein Damerau (default), etc.)}
}
\value{
Distance matrix between first and second BCR array determined by given distance.
}
\description{
\code{calculateDistances} returns the distance of the given arrays.
}
\details{
More Information here.
}
\examples{
#TODO


}
\seealso{
\code{\link[stringdist]{stringdist}}
}
\keyword{bcr}
\keyword{damerau}
\keyword{distance}
\keyword{levenshtein}
\keyword{metric}

